tests/test_core.py
import pytest
import pandas as pd
import numpy as np
import sys, os
sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))

from core.cleaning import clean_dataframe, parse_date_column
from core.analysis import compute_findings, safe_float


@pytest.fixture
def sample_df():
    return pd.DataFrame({
        "Order ID"    : ["CA-001", "CA-002", "CA-003", "CA-004"],
        "Order Date"  : ["11/08/2016", "11/09/2016", "06/15/2017", "09/20/2017"],
        "Ship Date"   : ["11/11/2016", "11/14/2016", "06/19/2017", "09/24/2017"],
        "Segment"     : ["Consumer", "Corporate", "Consumer", "Home Office"],
        "Region"      : ["West", "East", "West", "Central"],
        "Category"    : ["Furniture", "Technology", "Office Supplies", "Furniture"],
        "Sub-Category": ["Chairs", "Phones", "Paper", "Tables"],
        "Product ID"  : ["P001", "P002", "P003", "P004"],
        "Sales"       : [261.96, 731.94, 14.62, 957.58],
        "Quantity"    : [2, 3, 2, 5],
        "Discount"    : [0.0, 0.2, 0.0, 0.45],
        "Profit"      : [41.91, -219.58, 6.87, -383.03],
    })


def test_clean_returns_tuple(sample_df):
    result = clean_dataframe(sample_df.copy())
    assert isinstance(result, tuple)
    assert len(result) == 2


def test_columns_standardized(sample_df):
    df, _ = clean_dataframe(sample_df.copy())
    for col in df.columns:
        assert " " not in col
        assert "-" not in col


def test_sub_category_renamed(sample_df):
    df, _ = clean_dataframe(sample_df.copy())
    assert "Sub_Category" in df.columns


def test_dates_parsed(sample_df):
    df, _ = clean_dataframe(sample_df.copy())
    assert pd.api.types.is_datetime64_any_dtype(df["Order_Date"])


def test_campaign_flag(sample_df):
    df, _ = clean_dataframe(sample_df.copy())
    assert "Is_Campaign_Order" in df.columns
    # ✅ Fix: check flag matches original non-zero discount rows by index
    # rather than re-filtering on Discount after cleaning (outlier capping
    # may shift values, making exact > 0 comparisons unreliable)
    camp_count = df["Is_Campaign_Order"].sum()
    original_discount_count = (sample_df["Discount"] > 0).sum()
    assert camp_count == original_discount_count


def test_business_metrics(sample_df):
    df, _ = clean_dataframe(sample_df.copy())
    assert "Profit_Margin_"    in df.columns
    assert "Is_Campaign_Order" in df.columns
    assert "Discount_Band"     in df.columns


def test_findings_keys(sample_df):
    df, _ = clean_dataframe(sample_df.copy())
    f = compute_findings(df)
    for key in ["total_revenue", "total_profit", "overall_margin_%"]:
        assert key in f


def test_safe_float():
    assert safe_float(float("nan")) == 0.0
    assert safe_float(None)         == 0.0
    assert safe_float(42.5)         == 42.5


def test_date_parse_mm_dd_yyyy():
    s = pd.Series(["11/08/2016", "06/15/2017"])
    p = parse_date_column(s, "test")
    assert p.notna().all()
    assert pd.api.types.is_datetime64_any_dtype(p)


def test_delivery_days_positive(sample_df):
    df, _ = clean_dataframe(sample_df.copy())
    if "Delivery_Days" in df.columns:
        assert (df["Delivery_Days"].dropna() >= 0).all()