requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
google-generativeai==0.7.2
groq>=0.5.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0